codecrafters submit [skip ci]

+++ src/client/model.rs
@@ -5,7 +5,8 @@
 #[derive(Clone, Eq, PartialEq, Debug)]
 pub enum RespValue {
     SimpleString(String), // For Simple Strings the first byte of the reply is "+".
-    Integer(i64), // For Integers the first byte of the reply is ":".
+    // Integer(i64), // For Integers the first byte of the reply is ":".
+    Integer(u64), // For Integers the first byte of the reply is ":".
     BinaryBulkString(Vec<u8>), // For Bulk <binary> Strings the first byte of the reply is "$".
     BulkString(String), // For Bulk Strings the first byte of the reply is "$".
     Error(String), // For Errors the first byte of the reply is "-".
